Adding workaround for iOS8 bug.
The bug in question is that repeated taps to the alert controllers
dimming view causes repeated calls to ‘dismissViewController…’. This
means that the popover will dismiss after the first tap (as expected)
but then any presented view controllers will start dismissed. This
behaviour persists until the alert controller is fully dismissed.